// src/profile_utils/schema.ts
export interface ProfileObjectPermissions {
  object: string;
  allowCreate: boolean;
  allowDelete: boolean;
  allowEdit:boolean;
  allowRead:boolean;
  modifyAllRecords:boolean;
  viewAllRecords:boolean;
}
// ...
export interface ProfileUserPermission {
  name: string;
  enabled: boolean;
}